+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/exchangerate/widget/StyledExchangeUnavailablePopup.kt
@@ -1,42 +1,13 @@
 package co.electriccoin.zcash.ui.screen.exchangerate.widget
 
-import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.MutableTransitionState
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.GenericShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Popup
 import co.electriccoin.zcash.ui.R
-import co.electriccoin.zcash.ui.design.newcomponent.PreviewScreens
-import co.electriccoin.zcash.ui.design.theme.ZcashTheme
-import co.electriccoin.zcash.ui.design.theme.colors.ZashiColors
-import co.electriccoin.zcash.ui.design.theme.typography.ZashiTypography
+import co.electriccoin.zcash.ui.common.compose.ZashiAnimatedTooltip
+import co.electriccoin.zcash.ui.design.util.stringRes
 
 @Composable
 internal fun StyledExchangeUnavailablePopup(
@@ -49,88 +20,11 @@ internal fun StyledExchangeUnavailablePopup(
         onDismissRequest = onDismissRequest,
         offset = offset
     ) {
-        AnimatedVisibility(
+        ZashiAnimatedTooltip(
             visibleState = transitionState,
-            enter =
-                fadeIn() +
-                    slideInVertically(spring(stiffness = Spring.StiffnessHigh)) +
-                    scaleIn(spring(stiffness = Spring.StiffnessMedium, dampingRatio = Spring.DampingRatioLowBouncy)),
-            exit =
-                fadeOut() +
-                    scaleOut(animationSpec = spring(stiffness = Spring.StiffnessMedium)) +
-                    slideOutVertically(),
-        ) {
-            PopupContent(onDismissRequest = onDismissRequest)
-        }
-    }
-}
-
-@Suppress("MagicNumber")
-@Composable
-private fun PopupContent(onDismissRequest: () -> Unit) {
-    Column(
-        modifier = Modifier.padding(horizontal = 22.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Box(
-            modifier =
-                Modifier
-                    .width(16.dp)
-                    .height(8.dp)
-                    .background(ZashiColors.HintTooltips.surfacePrimary, TriangleShape)
+            title = stringRes(R.string.exchange_rate_unavailable_title),
+            message = stringRes(R.string.exchange_rate_unavailable_subtitle),
+            onDismissRequest = onDismissRequest
         )
-        Box(
-            Modifier
-                .fillMaxWidth()
-                .background(ZashiColors.HintTooltips.surfacePrimary, RoundedCornerShape(8.dp))
-                .padding(start = 12.dp, bottom = 12.dp),
-        ) {
-            Row {
-                Column(
-                    modifier = Modifier.weight(1f)
-                ) {
-                    Text(
-                        modifier = Modifier.padding(top = 12.dp),
-                        color = ZashiColors.Text.textLight,
-                        style = ZashiTypography.textMd,
-                        text = stringResource(R.string.exchange_rate_unavailable_title)
-                    )
-                    Spacer(modifier = Modifier.height(6.dp))
-                    Text(
-                        color = ZashiColors.Text.textLightSupport,
-                        style = ZashiTypography.textSm,
-                        text = stringResource(id = R.string.exchange_rate_unavailable_subtitle)
-                    )
-                }
-                IconButton(onClick = onDismissRequest) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_exchange_rate_unavailable_dialog_close),
-                        contentDescription = "",
-                        tint = ZashiColors.HintTooltips.defaultBg
-                    )
-                }
-            }
-        }
     }
 }
-
-private val TriangleShape =
-    GenericShape { size, _ ->
-
-        // 1) Start at the top center
-        moveTo(size.width / 2f, 0f)
-
-        // 2) Draw a line to the bottom right corner
-        lineTo(size.width, size.height)
-
-        // 3) Draw a line to the bottom left corner and implicitly close the shape
-        lineTo(0f, size.height)
-    }
-
-@Suppress("UnusedPrivateMember")
-@PreviewScreens
-@Composable
-private fun PopupContentPreview() =
-    ZcashTheme {
-        PopupContent(onDismissRequest = {})
-    }