+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/exportdata/view/ExportPrivateDataView.kt
@@ -1,52 +1,36 @@
 package co.electriccoin.zcash.ui.screen.exportdata.view
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
+import androidx.compose.ui.text.font.FontWeight
 import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
-import co.electriccoin.zcash.ui.design.MINIMAL_WEIGHT
-import co.electriccoin.zcash.ui.design.component.BlankBgScaffold
-import co.electriccoin.zcash.ui.design.component.Body
-import co.electriccoin.zcash.ui.design.component.LabeledCheckBox
-import co.electriccoin.zcash.ui.design.component.SmallTopAppBar
-import co.electriccoin.zcash.ui.design.component.TopAppBarBackNavigation
-import co.electriccoin.zcash.ui.design.component.TopScreenLogoTitle
 import co.electriccoin.zcash.ui.design.component.ZashiButton
+import co.electriccoin.zcash.ui.design.component.ZashiCheckbox
+import co.electriccoin.zcash.ui.design.component.ZashiSmallTopAppBar
+import co.electriccoin.zcash.ui.design.component.ZashiTopAppBarBackNavigation
+import co.electriccoin.zcash.ui.design.newcomponent.PreviewScreenSizes
 import co.electriccoin.zcash.ui.design.theme.ZcashTheme
+import co.electriccoin.zcash.ui.design.theme.colors.ZashiColors
+import co.electriccoin.zcash.ui.design.theme.dimensions.ZashiDimensions
+import co.electriccoin.zcash.ui.design.theme.typography.ZashiTypography
 import co.electriccoin.zcash.ui.design.util.scaffoldPadding
-
-@Preview("Export Private Data")
-@Composable
-private fun ExportPrivateDataPreview() {
-    ZcashTheme(forceDarkMode = false) {
-        ExportPrivateData(
-            snackbarHostState = SnackbarHostState(),
-            onBack = {},
-            onAgree = {},
-            onConfirm = {},
-            topAppBarSubTitleState = TopAppBarSubTitleState.None,
-        )
-    }
-}
+import co.electriccoin.zcash.ui.design.util.stringRes
 
 @Composable
 fun ExportPrivateData(
@@ -56,7 +40,7 @@ fun ExportPrivateData(
     onConfirm: () -> Unit,
     topAppBarSubTitleState: TopAppBarSubTitleState,
 ) {
-    BlankBgScaffold(
+    Scaffold(
         topBar = {
             ExportPrivateDataTopAppBar(
                 onBack = onBack,
@@ -82,19 +66,16 @@ private fun ExportPrivateDataTopAppBar(
     onBack: () -> Unit,
     subTitleState: TopAppBarSubTitleState
 ) {
-    SmallTopAppBar(
-        subTitle =
+    ZashiSmallTopAppBar(
+        title = stringResource(R.string.export_data_title),
+        subtitle =
             when (subTitleState) {
                 TopAppBarSubTitleState.Disconnected -> stringResource(id = R.string.disconnected_label)
                 TopAppBarSubTitleState.Restoring -> stringResource(id = R.string.restoring_wallet_label)
                 TopAppBarSubTitleState.None -> null
             },
         navigationAction = {
-            TopAppBarBackNavigation(
-                backText = stringResource(id = R.string.back_navigation).uppercase(),
-                backContentDescriptionText = stringResource(R.string.back_navigation_content_description),
-                onBack = onBack
-            )
+            ZashiTopAppBarBackNavigation(onBack = onBack)
         },
     )
 }
@@ -105,50 +86,35 @@ private fun ExportPrivateDataContent(
     onConfirm: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
-    Column(
-        modifier = modifier,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        TopScreenLogoTitle(
-            title = stringResource(R.string.export_data_header),
-            logoContentDescription = stringResource(R.string.zcash_logo_content_description)
+    Column(modifier = modifier) {
+        Text(
+            text = stringResource(R.string.export_data_header),
+            style = ZashiTypography.header6,
+            fontWeight = FontWeight.SemiBold,
+            color = ZashiColors.Text.textPrimary
         )
 
-        Spacer(Modifier.height(ZcashTheme.dimens.spacingLarge))
-
-        Body(
-            modifier = Modifier.testTag(ExportPrivateDataScreenTag.WARNING_TEXT_TAG),
-            text = stringResource(R.string.export_data_text_1)
-        )
-
-        Spacer(Modifier.height(ZcashTheme.dimens.spacingDefault))
+        Spacer(Modifier.height(ZashiDimensions.Spacing.spacingLg))
 
         Text(
-            modifier = Modifier.testTag(ExportPrivateDataScreenTag.ADDITIONAL_TEXT_TAG),
-            text = stringResource(R.string.export_data_text_2),
-            fontSize = 14.sp
+            modifier = Modifier.testTag(ExportPrivateDataScreenTag.WARNING_TEXT_TAG),
+            text = stringResource(R.string.export_data_text),
+            style = ZashiTypography.textSm,
+            color = ZashiColors.Text.textPrimary
         )
 
-        Spacer(Modifier.height(ZcashTheme.dimens.spacingDefault))
+        Spacer(Modifier.weight(1f))
 
         val checkedState = rememberSaveable { mutableStateOf(false) }
-        Row(Modifier.fillMaxWidth()) {
-            LabeledCheckBox(
-                checked = checkedState.value,
-                onCheckedChange = {
-                    checkedState.value = it
-                    onAgree(it)
-                },
-                text = stringResource(R.string.export_data_agree),
-                checkBoxTestTag = ExportPrivateDataScreenTag.AGREE_CHECKBOX_TAG
-            )
-        }
-
-        Spacer(
-            modifier =
-                Modifier
-                    .fillMaxHeight()
-                    .weight(MINIMAL_WEIGHT)
+        ZashiCheckbox(
+            modifier = Modifier.testTag(ExportPrivateDataScreenTag.AGREE_CHECKBOX_TAG),
+            isChecked = checkedState.value,
+            onClick = {
+                val new = checkedState.value.not()
+                checkedState.value = new
+                onAgree(new)
+            },
+            text = stringRes(R.string.export_data_agree),
         )
 
         Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingLarge))
@@ -161,3 +127,16 @@ private fun ExportPrivateDataContent(
         )
     }
 }
+
+@PreviewScreenSizes
+@Composable
+private fun ExportPrivateDataPreview() =
+    ZcashTheme {
+        ExportPrivateData(
+            snackbarHostState = SnackbarHostState(),
+            onBack = {},
+            onAgree = {},
+            onConfirm = {},
+            topAppBarSubTitleState = TopAppBarSubTitleState.None,
+        )
+    }