+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/seedrecovery/view/SeedRecoveryTag.kt
@@ -1,8 +1,0 @@
-package co.electriccoin.zcash.ui.screen.seedrecovery.view
-
-/**
- * These are only used for automated testing.
- */
-object SeedRecoveryTag {
-    const val DEBUG_MENU_TAG = "debug_menu"
-}