+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/whatsnew/model/WhatsNewState.kt
@@ -7,24 +7,28 @@ import co.electriccoin.zcash.ui.design.util.stringRes
 import kotlinx.datetime.LocalDate
 
 data class WhatsNewState(
-    val version: StringResource,
+    val titleVersion: StringResource,
+    val bottomVersion: StringResource,
     val date: LocalDate,
     val sections: List<WhatsNewSectionState>
 ) {
     companion object {
-        fun new(changelog: Changelog) =
-            WhatsNewState(
-                version = stringRes(R.string.whats_new_version, changelog.version),
-                date = changelog.date,
-                sections =
-                    listOfNotNull(changelog.added, changelog.changed, changelog.fixed, changelog.removed)
-                        .map {
-                            WhatsNewSectionState(
-                                stringRes(value = it.title),
-                                stringRes(it.content)
-                            )
-                        },
-            )
+        fun new(
+            changelog: Changelog,
+            version: String
+        ) = WhatsNewState(
+            titleVersion = stringRes(R.string.whats_new_version, changelog.version),
+            bottomVersion = stringRes(R.string.settings_version, version),
+            date = changelog.date,
+            sections =
+                listOfNotNull(changelog.added, changelog.changed, changelog.fixed, changelog.removed)
+                    .map {
+                        WhatsNewSectionState(
+                            stringRes(value = it.title),
+                            stringRes(it.content)
+                        )
+                    },
+        )
     }
 }