+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 - The in-app update logic has been fixed and is now correctly requested with every app launch
 - The Not enough space and In-app udpate screens have been redesigned
 - External links now open in in-app browser
+- All the Settings screens have been redesigned
 
 ### Fixed
 - Address book toast now correctly shows on send screen when adding both new and known addresses to text field